Fix parsed_from_(disc|cdstub) to actually work
It seems that as I was working on the last part I exclusively had
albums that didn't have cdstubs OR disc data, so I ended up not
finishing the work at all on this part.

While I was at it, I made the release counter start from 1 for a more
human-relatable experience. Of course, the editor is still vim, so,