Reshuffle Config, move dependencies around
No longer dependent on the external xdg module - it relied on
os.path.expandvars which, as the documentation notes, leaves nonexisting
variables unexpanded, caused it to start creating directories called
$HOME when the variable wasn't in the environment. While it is uncommon
it can occur during testing and we would rather the software crash in
that situation.

Fixed setup.py to not rely on being in the exact installation directory.
Fixed setup.py not looking for packages at all, and as such not finding
them (this seemed to be okay with pip install -e but it's a bit
inconvenient to ship empty packages).