Add test for failing to create config dir
This required a somewhat more involved process with the fixtures: Now we
have a few internal ones intended to setup the environment but they're
still exposed through similar ones as before to the actual tests.

Specifically, it seems we need to reload the modules between each test
function because otherwise we'll see stale values for the environment
variables which means tests will succeed or fail for reasons unrelated
to themselves. This is a bit unfortunate because it potentially involves
keeping track of what modules are imported by cdparacord.config...
recursively. However, that seems like a worst case that's not going to
happen just yet.

Also, remove the other ConfigError from creating directories. It
shouldn't happen, and if it does, we don't want it to be masked by
accident.