Start refactor in main.py
Basically, most things are to be moved to convenience classes, then
convenience functions, then a few are going to be left as-is.

Additionally, add vulture to dev dependencies. It's not automated in
any way because I don't think it's that useful to automate but I think
it's convenient to have a static dead code analyser.

This commit shouldn't work. It's ok to cherrypick master-compatible
changes and rebase this branch on that master, or to make those changes
directly in master.