Fix default configuration issue
The "parameters" map was a remnant from an old design and was masked by
local user configuration being used.

I don't know if we can unit test this particular thing but I can confirm
that the revised configuration works as it's identical to the LAME line
in  my personal local configuration.