Verify action configuration params are strings
First of all, we implement tests that ensure these conditions for the
default configuration:

1. Actions means encoder, as well as all the post_rip, post_encode and
   post_finished actions.
2. For each action, there must be exactly one dict key. Under this dict
   key must be a list containing only string values.

Then, we implement Dependency._verify_action_params which tests these
conditions for runtime config. Then, we implement tests for
_verify_action_params that take this into action.

In addition to this, the test fixtures in test_dependency have been
refactored to be slightly easier to use and to be more suitable for the
new tests. This required a minor change to most tests.