Both handle and test not getting track count
This is a bug but not a serious one - it's only important in the rare
case that somehow the disc is readable to discid but cdparanoia does not
correctly emit a table of contents. This seems most likely to occur if
it turns out `cdparanoia` is not actually a cdparanoia executable at
all.