Refactor _select_albumdata
It was, for some reason, checking whether the state was less than the
amount of tracks. This is strange, because it never depended on that,
and would have caused an erroneous output if there are more data sources
than tracks. This is not usual, but is very well possible for eg. an EP
or single with a lot of printings.

Additionally, instead of changing the elif to match against the length
of results, just rely on the fact that state starts at 0 and is bound to
its correct range [0, len(results)] when advanced. We do however add an
assert in case something ever changes this.