Separate safetyfilter default branch
Earlier it was elif safetyfilter != 'remove_restricted' which made the
test a bit strange to parse - technically it was complete, but I read it
wrong. I changed it to pass with an explanation and added an else for
the circumstance where no filters match.

The core reason for this change was that there's always an "else" branch
- if you don't define it you can imagine it as "else: pass". In this
case, the else branch would happen if and only if safetyfilter was
'remove_restricted' - but because that code would be hit anyway, and the
else branch was implicit and not explicit, it was confusing to me why I
needed a separate test for that case.

This variant where the else branch is explicit makes more sense from the
standpoint that all the if/elif branches are *valid* cases, and the
explicit else branch is an *invalid* case. In the earlier version, the
implicit else branch was also a valid case, which was hard to think of.