Change MusicBrainz fetching to support multi-CD
Now we don't ignore all the mediums beyond the first. Additionally,
multi-cd releases now keep track of which CD is in fact shown in the
results list.

Also, the warning on dropping sources has been downgraded to a note
(since it's not really your business, it's not dangerous, should only
happen in this case, and you really can't do anything about it anyway).