Move multi-artist check to albumdata constructor
This way we'll have the information carried along with the albumdata,
which will make its way into the rip process. Then, instead of checking
whether album and track artists match per-track, we'll simply check if
the album is multi-artist. This fulfills the intent of the
always_tag_albumartist setting which has been buggy thus far.

I'm personally affected by this bug relatively often but for some reason
I had not fixed it before.