Fix test_tag_track to use MutagenError
Earlier, the code this test tests caught any exception, and I had for
some reason chosen to raise a ValueError. In reality, it should only
catch MutagenError, so we need to raise a MutagenError to simulate the
condition of easyid3 failing.

This was a case of a correct fix in the code
breaking an incorrect test. Exciting.