Retype albumdata.py
It now uses TypedDicts to figure out the types of a lot of things.

First I wrote a partial typing for the outputs I get from MusicBrainz,
then I wrote a typing for the dict structures I internally pass around.
This allows a lot of cool things, like "I can finally typecheck the
dicts as something other than Dict[str, Any], because they always had a
schema".

The unfortunate part is that my own dicts have a schema that has a few
optional-but-actually-later-mandatory-they-just-get-filled-in-slowly
fields, which means it'll probably make more sense to make them
dataclasses with Optional fields, or something.

Basically the issue is that tracks don't immediately get track numbers
(Though I suppose they could) or filenames (because they haven't been
generated yet); and albumdata doesn't have a discid (because...
Actually, I don't know why?) or ripdir (because that only exists once we
start ripping).

In all likelihood in addition I should make an additional RipData thingy
that contains the extra information that none of the actual albumdata
inputs should have, but that is sometimes loadable. It's a bit of a mess.