Delete xdg module and fix tests
Basically, xdg.py has never been useful and has always been a problem.
So it's gone, now. All the tests in test_config.py have had their
fixtures changed to not be as much of a crime and a sin anymore.

Additionally, the old hack to get track count has been deleted - it
turns out that discid can give us the amount of audio tracks on a cd
just fine.