Update deps versions, add py3.8 testing
From now on I'm trying to remember to up the version whenever I make a
change after the previous version. The -dev versions are not required to
ever end up as an actual versions though.

PEP 440 specifies that version suffix ordering is:

    .devN, aN, bN, rcN, <no suffix>, .postN

and that `dev` is implicitly `dev0` so we should all fine with ordering
too.